cleanup and multithreading
