+++ src/main/resources/back_link_icon.txt
