+++ src/main/resources/file_icon.txt
