+++ src/main/resources/folder_icon.txt
